refactor: remove process utility file and integrate process definition directly in Vite config
