+++ portal/project-portal-common/src/index.ts
@@ -1,4 +1,4 @@
 export * from './utils';
 export * from './components';
 export * from './hooks';
-export * from "./process"
+
