--- portal/project-portal-common/src/process.ts
@@ -1,13 +1,0 @@
-const fix = `
-var process = {
-  env: {
-    NODE_ENV: "production"
-  }
-};
-var production = "production";
-`;
-
-export const InjectProcessPlugin = {
-  name: 'rollup-plugin-metadata',
-  renderChunk: (code: any) => fix + code,
-};
